services/sast_client.py
"""SAST Client for vulnerability scanning."""
import os
from typing import Dict, List, Optional
from google.cloud import aiplatform
import vertexai

class SASTClient:
    """Client for performing static application security testing."""

    def __init__(self, project_id: str, location: str = "us-central1"):
        """Initialize the SAST client.
        
        Args:
            project_id: The GCP project ID
            location: The GCP region where Vertex AI is deployed
        """
        self.project_id = project_id
        self.location = location
        
        # Initialize Vertex AI
        vertexai.init(project=project_id, location=location)
        self.ai_platform = aiplatform.init(project=project_id, location=location)

    def scan_code(self, repo_path: str) -> str:
        """Scan code for vulnerabilities using Vertex AI.
        
        Args:
            repo_path: Path to the repository to scan
            
        Returns:
            scan_id: Unique identifier for the scan
        """
        # TODO: Implement code scanning logic using Vertex AI
        model = aiplatform.Model.list(
            filter=f'display_name=sast-model',
            order_by='create_time desc',
            project=self.project_id,
            location=self.location
        )[0]

        # Create endpoint for inference
        endpoint = model.deploy()
        
        # Perform the scan
        # This is a placeholder - actual implementation would depend on your specific model
        return "scan_123"  # Return a scan ID

    def get_scan_results(self, scan_id: str) -> Dict:
        """Get the results of a security scan.
        
        Args:
            scan_id: The ID of the scan to retrieve results for
            
        Returns:
            Dict containing scan results
        """
        # TODO: Implement result retrieval logic
        # This would interact with your deployed model endpoint
        return {
            "scan_id": scan_id,
            "vulnerabilities": [],
            "status": "completed"
        }

    def analyze_vulnerabilities(self, scan_results: Dict) -> List[Dict]:
        """Analyze vulnerabilities found in scan results.
        
        Args:
            scan_results: The results from a security scan
            
        Returns:
            List of analyzed vulnerabilities with remediation steps
        """
        # TODO: Implement vulnerability analysis logic using Vertex AI
        endpoint = aiplatform.Endpoint(
            endpoint_name=f"projects/{self.project_id}/locations/{self.location}/endpoints/{scan_results['scan_id']}"
        )
        
        # This would be replaced with actual model inference
        return []

    def cleanup(self):
        """Cleanup resources."""
        # Implement cleanup logic for endpoints and other resources
        pass
